Fix failed install on new system
Dotbot plugins are stored as git submodules, so we need to make sure all
submodules are cloned before running dotbot. This will make sure that
dotbot can install fish and other required utilities.

Previously, the submodules were cloned by dotbot, but this meant that
they could not be used during the first run, and fish would be missing
when we tried to set it to the default shell.