Fix gitconfig not updating after initial install
Previously, the setup script would not make any updates to the gitconfig
if the customized file already exist. This file would exist after the
initial install when the name and email is put in, so any changes to the
template would not be automatically applied. Instead, the customized
file would have to be deleted and the dotfiles would have to be
installed again.

To fix this, I separated the custom part (author, email, etc) and the
rest of the gitconfig. If the author already exists, we dont need to
change it. But now, we always concatenate the author information and the
rest of the config file to update it.