Fix bufferline closing incorrect buffer
With the custom close function I had, if you had one buffer open but
pressed close on another buffer, it would close the buffer you were in
instead of the one you tried to close. This is because it was getting
the current buffer number instead of the one the close was requested on.