Fix double starship install on macOS
Starship is already installed in the script to install common packages,
so it is not needed in the macOS homebrew one.

I would prefer to have it installed by homebrew and apt instead of using
the install script, but the website doesn't mention an apt option.
Running the installer again should upgrade starship if necessary
according to their website docs.