Fix gitcommit files not having spellcheck enabled
I don't know if the neovim 0.8 changes caused some issues when there is
a lua and a vimscript file, but I just moved everything into the vim
file to fix the spell being disabled.

The change to treesitter is so that my theme can properly underline
misspelled words. With the regex option turned on, it seems to struggle
a bit.