Use my Go git-branches tool instead of the existing script
Python 3.11 broke the existing git-branches script because string enums
return their class name instead of their string value now which ruins
the output. Instead of worrying about fixing it, I am just going to
switch to the Go version of the tool that I have worked on for a little
bit. Right now it does basically the same thing, but I feel like the Go
version is a lot easier to maintain which is important as I plan to
expand it in the future to do more things.