Use custom Akoni font instead of Zed
I spent all the work to get a custom nerd font build and homebrew tap
for the Zed font before I realized that it does not have the terminal
spacing set to have the icons appear properly in the terminal. Instead,
I just made my own custom build of Iosevka and am going to start using
that instead. That way, I can more easily make changes to it in the
future, and I actually control the font.