Fix Go programs being installed to the default GOPATH
If the install script is run on a fresh system, we will have just
installed Go and won't have the GOPATH from my fish config loaded. This
causes the Go programs to be installed to ~/go, and when the user
launches the login shell with the correct GOPATH the programs wont work
since that default GOPATH wont be added to the PATH.

This reads the GOPATH setting from the fish config and exports that
before installing the programs with Go.