Fix configure-shell.sh not working on fresh install
If homebrew was just installed and used to install fish, then fish will
not be in the PATH. Sourcing the package install script allows it's
updates to PATH to be inherited by the main install script. That updated
PATH will then be passed to configure-shell.sh, allowing it to find
fish.