Fix incorrect environment in spotlight apps
When launching vscode on macOS with Spotlight or the dock, I noticed
that it wasn't able to find my installed tools, language servers, etc.
When I looked into it, I realized that it thought that zsh, macOS's
default login shell, was the shell it was supposed to use instead of
fish even though fish was the login shell. Because all of my paths and
other environment variables were set in config.fish, it had no idea
about them which explains why it couldnt find my tools.

To fix this, I'm now storing a POSIX-compliant profile in it's own file
and configuring multiple shells to be able to load it. This should
hopefully fix this issue when it happens in the future. For fish, the
fenv plugin is used to load the profile since fish doesnt understand the
export syntax by default.

Because of this change, the local override files have also been moved.
They now are in ~/.config/dots/ and there is a profile file (profile)
and a fish config (config.fish).