Fix docker check net working when docker is off
The `version` command will try to check the version of the server which
requires docker to be running. This check will only check the version of
the client which doesn't require it to be running. Previously, the check
would fail and start installing docker if docker was off but this should
fix that.