Fix install failures on low-spec machines
These plugins spin off a ton of download processes on a new install.
When the machine has low RAM or cores, this overwhelms the machine and
throws a ton of errors. Annoyingly, the errors that are spit out are
network-related (e.g. cannot resolve github.com) despite the connections
very clearly working.

The treesitter plugin doesn't seem to want to implement a config to
limit concurrency (very cool...) so instead we will just stick to
installing with sync instead.

Mason already maxes out it's concurrency at 4 which I think is fine.