Fix package install failing due to --no-lock option
The flag was deprecated/hidden a few months ago (seemingly because no
one actually wanted to use lock files for homebrew and thought they were
annoying, me too), and recently they completely removed it. This was
causing the install to fail with an error like 'Error: invalid option
--no-lock'. Since the files aren't generated anymore, the flag can just
be removed entirely.