Add command to disable formatting for future buffer saves
I have a command SaveWithoutFormatting which performs a single save
without formatting, but most of the time when I'm using this it's
because I'm in a file that has much different formatting then what my
local lsp or formatter wants. Because of that, I have to run that
command every single time I want to save to prevent the entire file from
being formatted. This approach allows me to disable formatting for an
entire buffer instead. Since this prevents the autocommand from
formatting the buffer, I can just save normally after running it.