Relax snacks.picker's aggressive truncation
When working on projects with more than a few nested folders, the
default truncation of 40 characters can make it pretty difficult to
figure out what file I'm looking at. For now, I'm just bumping it. I'd
like to have a function that changes the amount depending on how many
columns are available, but that's not a big deal.